Airflow/M3Capstone-Football_data_pipeline/requirements.txt
aiohttp==3.10.10
apache-airflow==2.10.0
loguru==0.7.2
pandas<2.2.0
python-dotenv==1.0.1
SQLAlchemy==1.4.54
dbt-core
dbt-postgres